redo file structure
